Create new module containing front page UI and server functions
The content of the app.R that builds the front page is moved to a separate module front_page.R.
The UI function contains the content boxes and actionLinks linking to the other navBar pages. The
server function creates a reactive value to be passed to an observer in app.R with the currently
selected navBar page.